# Conflicts:
#	src/formats/SipiIOJpeg.cpp

+++ include/SipiHttpServer.h
@@ -72,15 +72,17 @@ namespace Sipi {
         SipiHttpServer(int port_p, unsigned nthreads_p = 4, const std::string userid_str = "", const std::string &logfile_p = "sipi.log", const std::string &loglevel_p = "DEBUG");
         void run();
 
-        std::pair<std::string,std::string> get_canonical_url(int img_w,
-                                                             int img_h,
-                                                             const std::string &host,
-                                                             const std::string &prefix,
-                                                             const std::string &identifier,
-                                                             std::shared_ptr<SipiRegion> region,
-                                                             std::shared_ptr<SipiSize> size,
-                                                             SipiRotation &rotation,
-                                                             SipiQualityFormat &quality_format);
+        std::pair<std::string,std::string> get_canonical_url(
+            size_t img_w,
+            size_t img_h,
+            const std::string &host,
+            const std::string &prefix,
+            const std::string &identifier,
+            std::shared_ptr<SipiRegion> region,
+            std::shared_ptr<SipiSize> size,
+            SipiRotation &rotation,
+            SipiQualityFormat &quality_format
+        );
 
 
         inline pid_t pid(void) { return _pid; }
